multimodal-rag-windows/requirements.txt
langchain>=0.1.0
openai>=0.27.0
chromadb>=0.3.25
sentence-transformers>=2.2.2
transformers>=4.30.0
torch>=2.0.0
pdfplumber>=0.7.6
pymupdf>=1.22.5
pillow>=9.5.0
streamlit>=1.26.0
python-dotenv>=1.0.0
unstructured[local-inference]>=0.7.0
tiktoken>=0.4.0
sentencepiece>=0.1.99
